- Implementations of the Connection Protocol to Database Systems - Initial implementations of the ADO.NET embedded client

+++ src/deveeldb/Deveel.Data.Protocol/BeginResponse.cs
@@ -18,10 +18,10 @@ using System;
 namespace Deveel.Data.Protocol {
 	[Serializable]
 	public sealed class BeginResponse : IMessage {
-		public BeginResponse(int id) {
+		public BeginResponse(long id) {
 			Id = id;
 		}
 
-		public int Id { get; private set; }
+		public long Id { get; private set; }
 	}
 }
